requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.4.0=py311h06a4308_0
aiohttp=3.8.5=py311h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=4.2.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.2=py311h06a4308_0
attrs=23.1.0=py311h06a4308_0
babel=2.14.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=openblas
bleach=6.1.0=pypi_0
blinker=1.6.2=py311h06a4308_0
bottleneck=1.3.5=py311hbed6279_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
cairo=1.16.0=hb05425b_5
certifi=2024.2.2=py311h06a4308_0
cffi=1.15.1=py311h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py311h06a4308_0
comm=0.2.1=pypi_0
contourpy=1.0.5=py311hdb19cb5_0
cryptography=41.0.3=py311hdda0065_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
eigen=3.3.7=hd09550d_1
einops=0.7.0=pypi_0
executing=2.0.1=pypi_0
expat=2.5.0=h6a678d5_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.9.0=pypi_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py311h5eee18b_0
fsspec=2023.12.2=pypi_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=py311h06a4308_0
google-auth-oauthlib=0.5.2=py311h06a4308_0
graphite2=1.3.14=h295c915_1
grpc-cpp=1.48.2=he1ff14a_1
grpcio=1.48.2=py311he1ff14a_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
harfbuzz=4.3.0=hf52aaf7_1
hdf5=1.12.1=h2b7332f_3
huggingface-hub=0.20.2=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.29.0=pypi_0
ipython=8.20.0=pypi_0
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
jpeg=9e=h5eee18b_1
json5=0.9.14=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.1=pypi_0
jupyter-events=0.9.0=pypi_0
jupyter-lsp=2.2.2=pypi_0
jupyter-server=2.12.5=pypi_0
jupyter-server-terminals=0.5.2=pypi_0
jupyterlab=4.0.11=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.2=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.4=py311h6a678d5_0
kornia=0.7.1=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libcurl=8.4.0=h251f7ec_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libopenblas=0.3.21=h043d6bf_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py311h06a4308_0
markupsafe=2.1.1=py311h5eee18b_0
matplotlib=3.7.2=py311h06a4308_0
matplotlib-base=3.7.2=py311ha02d727_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py311h5eee18b_1
mpmath=1.3.0=pypi_0
multidict=6.0.2=py311h5eee18b_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
nbclient=0.9.0=pypi_0
nbconvert=7.14.2=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
notebook=7.0.7=pypi_0
notebook-shim=0.2.3=pypi_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.8.7=py311h812550d_0
numpy=1.26.0=py311h24aa872_0
numpy-base=1.26.0=py311hbfb1bba_0
oauthlib=3.2.2=py311h06a4308_0
opencv=4.6.0=py311h10ae9b0_5
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_1
overrides=7.6.0=pypi_0
packaging=23.1=py311h06a4308_0
pandas=2.1.1=py311ha02d727_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=h295c915_0
pexpect=4.9.0=pypi_0
pillow=10.0.1=py311ha6cbd5a_0
pip=23.2.1=py311h06a4308_0
pixman=0.40.0=h7f8727e_1
platformdirs=4.1.0=pypi_0
ply=3.11=py311h06a4308_0
prometheus-client=0.19.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=3.20.3=py311h6a678d5_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pypi_0
pyjwt=2.4.0=py311h06a4308_0
pyopenssl=23.2.0=py311h06a4308_0
pyparsing=3.0.9=py311h06a4308_0
pyqt=5.15.7=py311h6a678d5_0
pyqt5-sip=12.11.0=py311h6a678d5_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.5=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-json-logger=2.0.7=pypi_0
python-tzdata=2023.3=pyhd8ed1ab_0
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0=py311h5eee18b_1
pyzmq=25.1.2=pypi_0
qt-main=5.15.2=h7358343_9
qt-webengine=5.15.9=h9ab4d14_7
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
qtwebkit=5.212=h3fafdc1_5
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
referencing=0.32.1=pypi_0
requests=2.31.0=py311h06a4308_0
requests-oauthlib=1.3.0=py_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.17.1=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
safetensors=0.4.1=pypi_0
scipy=1.11.3=py311h24aa872_0
send2trash=1.8.2=pypi_0
setuptools=68.0.0=py311h06a4308_0
sip=6.6.2=py311h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
starfile=0.5.1=pyhd8ed1ab_0
sympy=1.12=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.12.1=py311h06a4308_0
tensorboard-data-server=0.7.0=py311h52d8a92_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.2=pyhd3eb1b0_0
terminado=0.18.0=pypi_0
thop=0.1.1-2209072238=pypi_0
timm=0.9.12=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torch=2.1.0+cu118=pypi_0
torchvision=0.16.0+cu118=pypi_0
tornado=6.3.3=py311h5eee18b_0
tqdm=4.65.0=py311h92b7b1e_0
traitlets=5.14.1=pypi_0
triton=2.1.0=pypi_0
types-python-dateutil=2.8.19.20240106=pypi_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
uri-template=1.3.0=pypi_0
urllib3=1.26.16=py311h06a4308_0
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
werkzeug=2.2.3=py311h06a4308_0
wheel=0.41.2=py311h06a4308_0
widgetsnbextension=4.0.9=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.2=h5eee18b_0
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py311h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
